requirements.txt
flask>=2.0.0
flask-wtf>=1.0.0
flask-bcrypt>=1.0.0
flask-limiter>=2.0.0
flask-sqlalchemy>=3.0.0
flask-migrate>=3.0.0
playsound==1.2.2
python-dotenv>=0.19.0
gunicorn>=20.1.0
email-validator>=1.3.0
blinker>=1.6.2